partly fixing tests, need to update selectors
